fix: build issue
자료형 잘못 적어서 빌드 안되는 문제 해결 및 소소한 리펙토링

+++ touhou_project/include/level/level.h
@@ -12,7 +12,7 @@
 
 class Level : public Base {
 private:
-  vector<Object *> layer_list_[LAYER_TYPE::END];
+  vector<Object *> layer_list_[static_cast<int>(LAYER_TYPE::END)];
   Player *current_player_;
 
 public:
